requirements.txt
# requirements.txt — Python dependencies for Litmus Lab (Zendesk training phase)

# python-dotenv — reads .env files and sets environment variables
python-dotenv==1.0.1

# requests — HTTP client for Zendesk REST API calls
requests>=2.31.0

# PyYAML — parses scenario YAML files
PyYAML>=6.0

# anthropic — Claude API SDK for grading trainee responses
anthropic>=0.40.0
